Add ability to write to file from logger (default is to stream)
daemonize now sets to write to logging files.

Move Daemon to daemon.py

Calling reload from daemon itself shouldn't require a pid file.
